SDR split into multiple files
Proxy moved to own file.
SDR-Metrics moved to own file.
Also split unit-tests into three files.
