+++ src/test/unit/ntypes/SdrMetricsTest.cpp
@@ -1,0 +1,540 @@
+/* ---------------------------------------------------------------------
+ * Copyright (C) 2019, David McDougall.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero Public License version 3 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Affero Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero Public License
+ * along with this program.  If not, see http://www.gnu.org/licenses.
+ * ----------------------------------------------------------------------
+ */
+
+#include <gtest/gtest.h>
+#include <nupic/ntypes/Sdr.hpp>
+#include <nupic/ntypes/SdrMetrics.hpp>
+/* Many of these tests also test SDR proxies. */
+#include <nupic/ntypes/SdrProxy.hpp>
+#include <vector>
+#include <random>
+
+using namespace std;
+using namespace nupic;
+
+/**
+ * SDR_Sparsity
+ * Test that it creates & destroys, and that nothing crashes.
+ */
+TEST(SdrMetrics, TestSparsityConstruct) {
+    SDR *A = new SDR({1});
+    SDR_Sparsity S( *A, 1000u );
+    ASSERT_ANY_THROW( SDR_Sparsity S( *A, 0u ) ); // Period > 0!
+    A->zero();
+    A->zero();
+    A->zero();
+    delete A; // Test use after freeing the parent SDR.
+    S.min();
+    S.max();
+    S.mean();
+    S.std();
+    ASSERT_EQ( S.sparsity, 0.0f );
+}
+
+TEST(SdrMetrics, TestSparsityExample) {
+    SDR A( { 1000u } );
+    SDR_Sparsity B( A, 1000u );
+    A.randomize( 0.01f );
+    A.randomize( 0.15f );
+    A.randomize( 0.05f );
+    ASSERT_EQ( B.sparsity, 0.05f);
+    ASSERT_EQ( B.min(),    0.01f);
+    ASSERT_EQ( B.max(),    0.15f);
+    ASSERT_NEAR( B.mean(),   0.07f, 0.005f );
+    ASSERT_NEAR( B.std(),    0.06f, 0.005f );
+}
+
+/*
+ * SDR_Sparsity
+ * Verify that the initial 10 values of metric are OK.
+ */
+TEST(SdrMetrics, TestSparsityShortTerm) {
+    SDR A({1});
+    SDR_Proxy B( A );
+    Real period = 10u;
+    Real alpha  = 1.0f / period;
+    SDR_Sparsity S( B, period );
+
+    A.setDense(SDR_dense_t{ 1 });
+    ASSERT_FLOAT_EQ( S.sparsity, 1.0f );
+    ASSERT_NEAR( S.min(),  1.0f, alpha );
+    ASSERT_NEAR( S.max(),  1.0f, alpha );
+    ASSERT_NEAR( S.mean(), 1.0f, alpha );
+    ASSERT_NEAR( S.std(),  0.0f, alpha );
+
+    A.setDense(SDR_dense_t{ 0 });
+    ASSERT_FLOAT_EQ( S.sparsity, 0.0f );
+    ASSERT_NEAR( S.min(),  0.0f, alpha );
+    ASSERT_NEAR( S.max(),  1.0f, alpha );
+    ASSERT_NEAR( S.mean(), 1.0f / 2, alpha );
+    ASSERT_NEAR( S.std(),  0.5f, alpha );
+
+    A.setDense(SDR_dense_t{ 0 });
+    ASSERT_FLOAT_EQ( S.sparsity, 0.0f );
+    ASSERT_NEAR( S.min(),  0.0f, alpha );
+    ASSERT_NEAR( S.max(),  1.0f, alpha );
+    ASSERT_NEAR( S.mean(), 1.0f / 3, alpha );
+    // Standard deviation was computed in python with numpy.std([ 1, 0, 0 ])
+    ASSERT_NEAR( S.std(),  0.47140452079103168f, alpha );
+
+    A.setDense(SDR_dense_t{ 0 });
+    ASSERT_NEAR( S.mean(), 1.0f / 4, alpha );
+    ASSERT_NEAR( S.std(),  0.4330127018922193f, alpha );
+
+    A.setDense(SDR_dense_t{ 0 });
+    ASSERT_NEAR( S.mean(), 1.0f / 5, alpha );
+    ASSERT_NEAR( S.std(),  0.40000000000000008f, alpha );
+
+    A.setDense(SDR_dense_t{ 0 });
+    ASSERT_NEAR( S.mean(), 1.0f / 6, alpha );
+    ASSERT_NEAR( S.std(),  0.372677996249965f, alpha );
+
+    A.setDense(SDR_dense_t{ 0 });
+    ASSERT_NEAR( S.mean(), 1.0f / 7, alpha );
+    ASSERT_NEAR( S.std(),  0.34992710611188266f, alpha );
+
+    A.setDense(SDR_dense_t{ 0 });
+    ASSERT_NEAR( S.mean(), 1.0f / 8, alpha );
+    ASSERT_NEAR( S.std(),  0.33071891388307384f, alpha );
+
+    A.setDense(SDR_dense_t{ 0 });
+    ASSERT_NEAR( S.mean(), 1.0f / 9, alpha );
+    ASSERT_NEAR( S.std(),  0.31426968052735443f, alpha );
+
+    A.setDense(SDR_dense_t{ 0 });
+    ASSERT_NEAR( S.mean(), 1.0f / 10, alpha );
+    ASSERT_NEAR( S.std(),  0.30000000000000004f, alpha );
+}
+
+/*
+ * SDR_Sparsity
+ * Verify that the longer run values of the SDR_Sparsity metric are OK.
+ * Test Protocol:
+ *      instantaneous-sparsity = Sample random distribution
+ *      for iteration in range( 1,000 ):
+ *          SDR.randomize( instantaneous-sparsity )
+ *      ASSERT_NEAR( SparsityMetric.mean(), true_mean )
+ *      ASSERT_NEAR( SparsityMetric.std(),  true_std )
+ */
+TEST(SdrMetrics, TestSparsityLongTerm) {
+    auto period     = 100u;
+    auto iterations = 1000u;
+
+    SDR A({1000u});
+    SDR_Proxy B( A );
+    SDR_Sparsity S( B, period );
+
+    vector<Real> test_means{ 0.01f,  0.05f,  0.20f, 0.50f, 0.50f, 0.75f, 0.99f };
+    vector<Real> test_stdev{ 0.001f, 0.025f, 0.10f, 0.33f, 0.01f, 0.15f, 0.01f };
+
+    std::default_random_engine generator;
+    for(auto test = 0u; test < test_means.size(); test++) {
+        const auto mean = test_means[test];
+        const auto stdv = test_stdev[test];
+        auto dist = std::normal_distribution<float>(mean, stdv);
+        for(UInt i = 0; i < iterations; i++) {
+            Real sparsity;
+            do {
+                sparsity = dist( generator );
+            } while( sparsity < 0.0f || sparsity > 1.0f);
+            A.randomize( sparsity );
+            EXPECT_NEAR( S.sparsity, sparsity, 0.501f / A.size );
+        }
+        EXPECT_NEAR( S.mean(), mean, stdv );
+        EXPECT_NEAR( S.std(),  stdv, stdv / 2.0f );
+    }
+}
+
+TEST(SdrMetrics, TestSparsityPrint) {
+    // TODO: Automatically test.  Use regex or other parsing utility.  Extract
+    // the data into real numbers and check that its acceptably near to the
+    // expected.
+    cerr << endl << "YOU must manually verify this output!" << endl << endl;
+    SDR A({ 2000u });
+    SDR_Proxy B( A );
+    SDR_Sparsity S( B, 10u );
+
+    A.randomize( 0.30f );
+    A.randomize( 0.70f );
+    cerr << S;
+
+    A.randomize( 0.123456789f );
+    A.randomize( 1.0f - 0.123456789f );
+    cerr << S;
+    cerr << endl;
+}
+
+/**
+ * SDR_ActivationFrequency
+ * Test that it creates & destroys, and that no methods crash.
+ */
+TEST(SdrMetrics, TestAF_Construct) {
+    // Test creating it.
+    SDR *A = new SDR({ 5 });
+    SDR_Proxy *B = new SDR_Proxy( *A );
+    SDR_ActivationFrequency F( *B, 100 );
+    ASSERT_ANY_THROW( SDR_ActivationFrequency F( *A, 0u ) ); // Period > 0!
+    // Test nothing crashes with no data.
+    F.min();
+    F.mean();
+    F.std();
+    F.max();
+    ASSERT_EQ( F.activationFrequency.size(), A->size );
+
+    // Test with junk data.
+    A->zero(); A->randomize( 0.5f ); A->randomize( 1.0f ); A->randomize( 0.5f );
+    F.min();
+    F.mean();
+    F.std();
+    F.max();
+    ASSERT_EQ( F.activationFrequency.size(), A->size );
+
+    // Test use after freeing parent SDR.
+    auto A_size = A->size;
+    delete A;
+    delete B;
+    F.min();
+    F.mean();
+    F.std();
+    F.max();
+    ASSERT_EQ( F.activationFrequency.size(), A_size );
+}
+
+/**
+ * SDR_ActivationFrequency
+ * Verify that the first few data points are ok.
+ */
+TEST(SdrMetrics, TestAF_Example) {
+    SDR A({ 2u });
+    SDR_Proxy B( A );
+    SDR_ActivationFrequency F( B, 10u );
+
+    A.setDense(SDR_dense_t{ 0, 0 });
+    ASSERT_EQ( F.activationFrequency, vector<Real>({ 0.0f, 0.0f }));
+
+    A.setDense(SDR_dense_t{ 1, 1 });
+    ASSERT_EQ( F.activationFrequency, vector<Real>({ 0.5f, 0.5f }));
+
+    A.setDense(SDR_dense_t{ 0, 1 });
+    ASSERT_NEAR( F.activationFrequency[0], 0.3333333333333333f, 0.001f );
+    ASSERT_NEAR( F.activationFrequency[1], 0.6666666666666666f, 0.001f );
+    ASSERT_EQ( F.min(), F.activationFrequency[0] );
+    ASSERT_EQ( F.max(), F.activationFrequency[1] );
+    ASSERT_FLOAT_EQ( F.mean(), 0.5f );
+    ASSERT_NEAR( F.std(), 0.16666666666666666f, 0.001f );
+    ASSERT_NEAR( F.entropy(), 0.9182958340544896f, 0.001f );
+}
+
+/*
+ * SDR_ActivationFrequency
+ * Verify that the longer run values of this metric are OK.
+ */
+TEST(SdrMetrics, TestAF_LongTerm) {
+    const auto period  =   100u;
+    const auto runtime = 10000u;
+    SDR A({ 20u });
+    SDR_Proxy B( A );
+    SDR_ActivationFrequency F( B, period );
+
+    vector<Real> test_sparsity{ 0.0f, 0.02f, 0.05, 1.0f, 0.25f, 0.5f };
+
+    for(const auto &sparsity : test_sparsity) {
+        for(UInt i = 0; i < runtime; i++)
+            A.randomize( sparsity );
+
+        const auto epsilon = 0.10f;
+        EXPECT_GT( F.min(), sparsity - epsilon );
+        EXPECT_LT( F.max(), sparsity + epsilon );
+        EXPECT_NEAR( F.mean(), sparsity, epsilon );
+        EXPECT_NEAR( F.std(),  0.0f,     epsilon );
+    }
+}
+
+TEST(SdrMetrics, TestAF_Entropy) {
+    const auto size    = 1000u; // Num bits in SDR.
+    const auto period  =  100u; // For activation frequency exp-rolling-avg
+    const auto runtime = 1000u; // Train time for each scenario
+    const auto tolerance = 0.02f;
+
+    // Extact tests:
+    // Test all zeros.
+    SDR A({ size });
+    SDR_Proxy Px( A );
+    SDR_ActivationFrequency F( Px, period );
+    A.zero();
+    EXPECT_FLOAT_EQ( F.entropy(), 0.0f );
+
+    // Test all ones.
+    SDR B({ size });
+    SDR_ActivationFrequency G( B, period );
+    B.randomize( 1.0f );
+    EXPECT_FLOAT_EQ( G.entropy(), 0.0f );
+
+    // Probabilistic tests:
+    // Start with random SDRs, verify 100% entropy. Then disable cells and
+    // verify that the entropy decreases.  Disable cells by freezing their value
+    // so that the resulting SDR keeps the same sparsity.  Progresively disable
+    // more cells and verify that the entropy monotonically decreases.  Finally
+    // verify 0% entropy when all cells are disabled.
+    SDR C({ size });
+    SDR_ActivationFrequency H( C, period );
+    auto last_entropy = -1.0f;
+    const UInt incr = size / 10u; // NOTE: This MUST divide perfectly, with no remainder!
+    for(auto nbits_disabled = 0u; nbits_disabled <= size; nbits_disabled += incr) {
+        for(auto i = 0u; i < runtime; i++) {
+            SDR scratch({size});
+            scratch.randomize( 0.05f );
+            // Freeze bits, such that nbits remain alive.
+            for(auto z = 0u; z < nbits_disabled; z++)
+                scratch.getDense()[z] = C.getDense()[z];
+            C.setDense( scratch.getDense() );
+        }
+        const auto entropy = H.entropy();
+        if( nbits_disabled == 0u ) {
+            // Expect 100% entropy
+            EXPECT_GT( entropy, 1.0f - tolerance );
+        }
+        else{
+            // Expect less entropy than last time, when fewer bits were disabled.
+            ASSERT_LT( entropy, last_entropy );
+        }
+        last_entropy = entropy;
+    }
+    // Expect 0% entropy.
+    EXPECT_LT( last_entropy, tolerance );
+}
+
+TEST(SdrMetrics, TestAF_Print) {
+    // TODO: Automatically test.  Use regex or other parsing utility.  Extract
+    // the data into real numbers and check that its acceptably near to the
+    // expected.
+    const auto period  =  100u;
+    const auto runtime = 1000u;
+    cerr << endl << "YOU must manually verify this output!" << endl << endl;
+    SDR A({ 2000u });
+    SDR_ActivationFrequency F( A, period );
+
+    vector<Real> sparsity{ 0.0f, 0.02f, 0.05f, 0.50f, 0.0f };
+    for(const auto sp : sparsity) {
+        for(auto i = 0u; i < runtime; i++)
+            A.randomize( sp );
+        cerr << F;
+        cerr << endl;
+    }
+}
+
+TEST(SdrMetrics, TestOverlap_Construct) {
+    SDR *A = new SDR({ 1000u });
+    SDR_Overlap V( *A, 100u );
+    ASSERT_ANY_THROW( new SDR_Overlap( *A, 0 ) ); // Period > 0!
+    // Check that it doesn't crash, when uninitialized.
+    V.min();
+    V.mean();
+    V.std();
+    V.max();
+    // If no data, have obviously wrong result.
+    ASSERT_FALSE( V.overlap >= 0.0f and V.overlap <= 1.0f );
+
+    // Check that it doesn't crash with half enough data.
+    A->randomize( 0.20f );
+    V.min();
+    V.mean();
+    V.std();
+    V.max();
+    ASSERT_FALSE( V.overlap >= 0.0f and V.overlap <= 1.0f );
+
+    // Check no crash with data.
+    A->addNoise( 0.50f );
+    V.min();
+    V.mean();
+    V.std();
+    V.max();
+    ASSERT_EQ( V.overlap, 0.50f );
+
+    // Check overlap metric is valid after parent SDR is deleted.
+    delete A;
+    V.min();
+    V.mean();
+    V.std();
+    V.max();
+    ASSERT_EQ( V.overlap, 0.50f );
+}
+
+TEST(SdrMetrics, TestOverlap_Example) {
+    SDR A({ 10000u });
+    SDR_Proxy Px( A );
+    SDR_Overlap B( Px, 1000u );
+    A.randomize( 0.05 );
+    A.addNoise( 0.95 );         //   5% overlap
+    A.addNoise( 0.55 );         //  45% overlap
+    A.addNoise( 0.72 );         //  28% overlap
+    ASSERT_EQ( B.overlap,  0.28f );
+    ASSERT_EQ( B.min(),    0.05f );
+    ASSERT_EQ( B.max(),    0.45f );
+    ASSERT_NEAR( B.mean(), 0.26f, 0.005f );
+    ASSERT_NEAR( B.std(),  0.16f, 0.005f );
+}
+
+TEST(SdrMetrics, TestOverlap_ShortTerm) {
+    SDR         A({ 1000u });
+    SDR_Proxy   Px( A );
+    SDR_Overlap V( Px, 10u );
+
+    A.randomize( 0.20f ); // Initial value is taken after SDR_Overlap is created
+
+    // Add overlap 50% to metric tracker.
+    A.addNoise(  0.50f );
+    ASSERT_FLOAT_EQ( V.overlap, 0.50f );
+    ASSERT_FLOAT_EQ( V.min(),   0.50f );
+    ASSERT_FLOAT_EQ( V.max(),   0.50f );
+    ASSERT_FLOAT_EQ( V.mean(),  0.50f );
+    ASSERT_FLOAT_EQ( V.std(),   0.0f );
+
+    // Add overlap 80% to metric tracker.
+    A.addNoise(  0.20f );
+    ASSERT_FLOAT_EQ( V.overlap, 0.80f );
+    ASSERT_FLOAT_EQ( V.min(),   0.50f );
+    ASSERT_FLOAT_EQ( V.max(),   0.80f );
+    ASSERT_FLOAT_EQ( V.mean(),  0.65f );
+    ASSERT_FLOAT_EQ( V.std(),   0.15f );
+
+    // Add overlap 25% to metric tracker.
+    A.addNoise(  0.75f );
+    ASSERT_FLOAT_EQ( V.overlap, 0.25f );
+    ASSERT_FLOAT_EQ( V.min(),   0.25f );
+    ASSERT_FLOAT_EQ( V.max(),   0.80f );
+    ASSERT_FLOAT_EQ( V.mean(),  0.51666666666666672f ); // Source: python numpy.mean
+    ASSERT_FLOAT_EQ( V.std(),   0.22484562605386735f ); // Source: python numpy.std
+}
+
+TEST(SdrMetrics, TestOverlap_LongTerm) {
+    const auto runtime = 1000u;
+    const auto period  =  100u;
+    SDR A({ 500u });
+    SDR_Proxy Px( A );
+    SDR_Overlap V( Px, period );
+    A.randomize( 0.45f );
+
+    vector<Real> mean_ovlp{ 0.0f, 1.0f,
+                            0.5f, 0.25f,
+                            0.85f, 0.95f };
+
+    vector<Real> std_ovlp{  0.01f, 0.01f,
+                            0.33f, 0.05f,
+                            0.05f, 0.02f };
+
+    std::default_random_engine generator;
+    for(auto i = 0u; i < mean_ovlp.size(); i++) {
+        auto dist = std::normal_distribution<float>(mean_ovlp[i], std_ovlp[i]);
+
+        for(auto z = 0u; z < runtime; z++) {
+            Real ovlp;
+            do {
+                ovlp = dist( generator );
+            } while( ovlp < 0.0f || ovlp > 1.0f );
+            A.addNoise( 1.0f - ovlp );
+            EXPECT_NEAR( V.overlap, ovlp, 0.501f / A.getSum() );
+        }
+        EXPECT_NEAR( V.mean(), mean_ovlp[i], std_ovlp[i] );
+        EXPECT_NEAR( V.std(),  std_ovlp[i],  std_ovlp[i] / 2.0f );
+    }
+}
+
+TEST(SdrMetrics, TestOverlap_Print) {
+    // TODO: Automatically test.  Use regex or other parsing utility.  Extract
+    // the data into real numbers and check that its acceptably near to the
+    // expected.
+    cerr << endl << "YOU must manually verify this output!" << endl << endl;
+    SDR A({ 2000u });
+    SDR_Proxy Px( A );
+    SDR_Overlap V( Px, 100u );
+    A.randomize( 0.02f );
+
+    vector<Real> overlaps{ 0.02f, 0.05f, 0.0f, 0.50f, 0.0f };
+    for(const auto ovlp : overlaps) {
+        for(auto i = 0u; i < 1000u; i++)
+            A.addNoise( 1.0f - ovlp );
+        cerr << V;
+    }
+    for(auto i = 0u; i < 1000u; i++)
+        A.randomize( 0.02f );
+    cerr << V;
+    cerr << endl;
+}
+
+/**
+ * SDR_Metrics
+ *
+ */
+TEST(SdrMetrics, TestAllMetrics_Construct) {
+    // Test that it constructs.
+    SDR *A = new SDR({ 100u });
+    SDR_Metrics M( *A, 10u );
+
+    A->randomize( 0.05f );
+    A->randomize( 0.05f );
+    A->randomize( 0.05f );
+
+    // Test use after freeing data source.
+    delete A;
+    stringstream devnull;
+    devnull << M;
+
+    // Test delete Metric and keep using SDR.
+    A = new SDR({ 100u });
+    SDR_Metrics *B = new SDR_Metrics( *A, 99u );
+    SDR_Metrics *C = new SDR_Metrics( *A, 98u );
+    A->randomize( 0.20f );
+    A->randomize( 0.20f );
+    devnull << *B;
+    delete B;                   // First deletion
+    A->randomize( 0.20f );
+    A->addNoise( 0.20f );
+    B = new SDR_Metrics( *A, 99u );    // Remove & Recreate
+    A->randomize( 0.20f );
+    A->randomize( 0.20f );
+    devnull << *A;
+    delete A;
+    devnull << *C;
+    delete C;
+    devnull << *B;
+    delete B;
+}
+
+/**
+ * SDR_Metrics prints OK.
+ */
+TEST(SdrMetrics, TestAllMetrics_Print) {
+    // TODO: Automatically test.  Use regex or other parsing utility.  Extract
+    // the data into real numbers and check that its acceptably near to the
+    // expected.
+    cerr << endl << "YOU must manually verify this output!" << endl << endl;
+    SDR A({ 4097u });
+    SDR_Proxy Px( A );
+    SDR_Metrics M( Px, 100u );
+
+    vector<Real> sparsity{ 0.02f, 0.15f, 0.06f, 0.50f, 0.0f };
+    vector<Real> overlaps{ 0.02f, 0.05f, 0.10f, 0.50f, 0.0f };
+    for(auto test = 0u; test < sparsity.size(); test++) {
+        A.randomize( sparsity[test] );
+        for(auto i = 0u; i < 1000u; i++)
+            A.addNoise( 1.0f - overlaps[test] );
+        cerr << M;
+        cerr << endl;
+    }
+}
+
